dialogo de confirmação

+++ src/app/courses/containers/courses/courses.component.ts
@@ -8,6 +8,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -71,19 +72,29 @@ export class CoursesComponent implements OnInit {
   }
 
   onDelete(course: Course) {
-    this.coursesService.remove(course._id).subscribe( //sem o subscribe nao vai fazer a requisição
-      () => {
-        this.refresh();
-        this.snackBar.open('Curso removido com sucesso!', 'X', {
-          duration: 2000,
-          verticalPosition: 'top', horizontalPosition: 'center'
-        },);
-      },
-      error => {
-        this.onError('Não foi possível remover o curso');
-      }
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: `Deseja realmente remover o curso ${course.name}?`,
+    });
 
-    )
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe( //sem o subscribe nao vai fazer a requisição
+          () => {
+            this.refresh();
+            this.snackBar.open('Curso removido com sucesso!', 'X', {
+              duration: 2000,
+              verticalPosition: 'top', horizontalPosition: 'center'
+            },);
+          },
+          error => {
+            this.onError('Não foi possível remover o curso');
+          }
+
+        );
+      }
+    });
+
+
   }
 
 }
